requirements.txt
streamlit
langchain
python-dotenv
langchain-pinecone
langchain-openai


langsmith
langchain-community
tiktoken
langchainhub
langgraph